+++ frontend/src/components/Header.jsx
@@ -14,6 +14,7 @@ import {
 } from "@mantine/core";
 import { Sun, Moon, DotsVertical } from "tabler-icons-react";
 import "@mantine/core/styles.css";
+import Register from "./Register";
 
 const HeaderComponent = () => {
   const [darkMode, setDarkMode] = useState(false);
@@ -50,11 +51,7 @@ const HeaderComponent = () => {
     setIsLoggedIn(false);
   };
 
-  const handleRegisterSubmit = (e) => {
-    e.preventDefault();
-    console.log("Registrierung abgeschlossen!");
-    setShowRegisterMenu(false); // Schließt das Registrierungsmenü
-  };
+  
 
   return (
     <div style={{ height: 70, padding: "0 16px", backgroundColor: "lightblue" }}>
@@ -131,62 +128,7 @@ const HeaderComponent = () => {
       </Container>
 
       {/* Registration Menu */}
-      {showRegisterMenu && (
-  <>
-    <Overlay opacity={0.6} color="#000" zIndex={1000} />
-    <Paper
-      shadow="xl"
-      radius="md"
-      p="xl"
-      style={{
-        position: "fixed",
-        top: "50%",
-        left: "50%",
-        transform: "translate(-50%, -50%)",
-        zIndex: 1100,
-        width: "90%",
-        maxWidth: "400px",
-      }}
-    >
-      <Title order={3} align="center" mb="lg">
-        Registrieren
-      </Title>
-      <form onSubmit={handleRegisterSubmit}>
-        <TextInput
-          label="Name"
-          placeholder="Dein Name"
-          name="name"
-          required
-          mb="sm"
-        />
-        <TextInput
-          label="E-Mail"
-          placeholder="Deine E-Mail-Adresse"
-          type="email"
-          name="email"
-          required
-          mb="sm"
-        />
-        <PasswordInput
-          label="Passwort"
-          placeholder="Dein Passwort"
-          name="password"
-          required
-          mb="sm"
-        />
-        <Group position="apart" mt="xl">
-          <Button
-            variant="default"
-            onClick={() => setShowRegisterMenu(false)}
-          >
-            Abbrechen
-          </Button>
-          <Button type="submit">Registrieren</Button>
-        </Group>
-      </form>
-    </Paper>
-  </>
-)}
+      {showRegisterMenu && (<Register />)}
     </div>
   );
 };
